updates
staging examples to be capable with master
