--- examples/helloworld/go.mod
@@ -1,3 +1,0 @@
-module helloworld
-
-go 1.18
