--- examples/middlewares/go.mod
@@ -1,3 +1,0 @@
-module middlewares
-
-go 1.18
