--- examples/pathgroups/go.mod
@@ -1,3 +1,0 @@
-module pathgroups
-
-go 1.18
