--- examples/requestbody/go.mod
@@ -1,3 +1,0 @@
-module requestbody
-
-go 1.18
